Add a new authenticator for opaque access tokens
This authenticator authenticates requests that include an opaque access
token in their Authorization header. AuthService performs a request to
the `/userinfo` endpoint of the integrated external Identity Provider.
The successful response of the provider:
* validates that the opaque access token is valid and not expired
* returns a series of claims related to the user

AuthService retrieves the USERID_CLAIM and the GROUPS_CLAIM of the
Provider's response, it returns HTTP 200 status and set the UserID
header for the client.

GitHub-PR: #

Signed-off-by: Athanasios Markou <athamark@arrikto.com>