settings: introduce the ACCESS_TOKEN_AUTHN envvar
Remove the JWT_AUTHN_ENABLED configuration and add a new
configuration via the ACCESS_TOKEN_AUTHN env var. This configuration
accepts values of type string and specifically:
- "JWT": enables the JWT authenticator and disables the opaque
  authenticator
- "opaque": enables the opaque authenticator and disables the JWT
  authenticator
- "none": disables both access token authenticators
Any deviations from the above, result in an Error log message that
indicates that this environment configuration was not set properly
by the admin.

GitHub-PR: #

Signed-off-by: AthanasiosMarkou <athamark@arrikto.com>